drop C++17 filesystem features for backward compatibility for openSUSE Leap
It turns out that Leap 15.3 is still stuck with gcc 7.5 which doesn't
fully implement C++17. The filesystem header is still in experimental
state there and doesn't provide everything we need.

There exist newer gcc versions that explicitly need to be pulled in
explicitly, but this could give us trouble, because permissions are part
of all OBS package build requirements.

Therefore adjust the code to not require these newer features to become
compatible with gcc 7.5 again.