profiles: remove extensive documentation from files in favor of man page
There is a lot of outdated and confusing documentation in these file
headers. We have a common man page for the configuration files which
avoids the redundancy we see here in these files. Also these files are
not supposed to act as a kind of hardening guide, there are better
sources for this.

Therefore reduce the documentation to a minimum, referring to the man
page and describing the intended purpose of the profile in question.