EntryProcessor: enforce package ownership limitation
Only operate on files on disk if they are actually owned by the expected
package(s).

To this end we need to be able to invoke child processes for the first
time (an alternative would have been to link against librpm or something
like that, but that would be even more complex, I fear). Thus this
change introduces a new helper function which goes through all the
mumbo-jumbo of child process creation, pipe setup etc.

A couple of new error situations can result from this, when child
process setup fails, or `rpm` fails. We try to be conservative and
assume no package name match in error situations.